// src/data/nightfarers/types.ts
export type Nightfarer = {
  index: number;
  name: string;
  skills: string;
  skill1: string;
  skill2: string;
  skill3?: string;
  description: string;
  description1: string;
  description2: string;
  equipment: string;
  equipment1: string;
  equipment2: string;
  nameImageUrl: string;
  skillImageUrls: string;
  equipmentImageUrls: string;
  rawText: string;
  aboutText: string;
};